Adjusted customer put method

+++ src/main/java/com/diogo/barbernet/api/services/ClienteService.java
@@ -1,5 +1,6 @@
 package com.diogo.barbernet.api.services;
 
+import com.diogo.barbernet.api.domain.ValidacaoException;
 import com.diogo.barbernet.api.domain.cliente.Cliente;
 import com.diogo.barbernet.api.domain.cliente.ClienteRepository;
 import com.diogo.barbernet.api.domain.cliente.DadosAtulizacaoCliente;
@@ -20,15 +21,18 @@ public class ClienteService {
         Optional<Cliente> optionalCliente = repository.findById(id);
         return optionalCliente.orElse(null);
     }
+
     public Cliente cadastrarCliente(DadosCadastroCliente dados){
         Cliente newDados = new Cliente(dados);
         return repository.save(newDados);
     }
+
     public List<Cliente> findAll() {
         return repository.findAll();
     }
-    public Cliente atualizarCliente(DadosAtulizacaoCliente dados) {
-        Cliente cliente = findById(dados.id());
+
+    public Cliente atualizarCliente(Long id, DadosAtulizacaoCliente dados) {
+        Cliente cliente = findById(id);
         cliente.setNome(dados.nome());
         cliente.setTelefone(dados.telefone());
         cliente.setEmail(dados.email());
@@ -36,8 +40,12 @@ public class ClienteService {
 
         return repository.save(cliente);
     }
+
     public void deletar(Long id) {
         Cliente cliente = findById(id);
+        if(cliente.getAgendamentos().size() > 0) {
+            throw new ValidacaoException("cliente possui agendamentos");
+        }
         repository.deleteById(id);
     }
 }
